indent/nim: fix autoindent
Allows overriding paren-aware indent with autoindent. This is done by
only use the current parenthesis indent when it's of the line before the
current, then (ab)use autoindent to keep everything inline.

    tuple[a: int,
          b: int,
          ^

By employing autoindent, the user is also allowed to override the
suggested indent should they disagree with it.