suggest/[manager, utils]: refactor querying interface, pt. 1
The querying interface is now a part of suggest/manager.vim. This
features usage of exception for error message passing, reducing repeated
code.

A much cleaner querying interface is also a part of this refactor. See
the docs for more information.

The nim#suggest#utils#Query() interface is currently being kept until
everything is mitigated to the new interface