suggest/manager: isolate user-passed options
This provides a complete layer of abstraction on top of the editor's native
tcp communication infrastructure.

While the old "sockconnect() and chansend() wrapper but not really" way
allow for maximum customizability, it'll be a burden for the API users to
keep track of whether they've passed any reserved key, especially when
multiple different editors are going to be supported by the plugin.