syntax/nim: restrict non-raw long string sync point to common idioms
The old way was too inclusive and causes wrong highlighting in case
the long string ends on the same line. This happens because long non-raw
strings have the same starting point as the ending point.

Instead we try to infer possible starting points by making use of
certain code structures that these strings might be a part of. So far I
found `discard` and `asm` to be the most prominent.