autoload/*: correctness and micro optimization
We swap most `==` and `!=` into `is[#?]` and `isnot[#?]`. This is
because using `==` actually causes undefined behavior!

`==` depends on the user's ignorecase, to portably compare we need to
use their `#` or `?` variant instead, to take casing into account. This
also let us micro optimize a tiny bit by specifying `#` for
case-sensitive matches, will in theory takes less time.

The switch to `is`/`isnot` is more of a micro optimization than
correctness. By default `==` will try to convert the arguments into
the same type, so we use `is` instead, which don't have this behavior.
Most of the codebase is rather correctly typed, so this won't change
much. However, it will let us skip potential type conversions.

This change don't bring us any observable speed up, but it's something.