suggest/manager: use full paths for query
Apparently bufname() don't return the full path as I've thought. This
causes nimsuggest to not perform the correct operations on correct files
if nimsuggest wasn't spawned on the same location as the current
directory in vim.

This bug was found by @mfiano. It probably went undetected since most
don't do `:cd` in neovim.