indent/nim: the colon trigger returns!
`except` is commonly used enough that triggering dedent on this is
usually harmful.

Instead we bring back the `:` as a trigger, which has now been limited
to only activate at the end of line and can only trigger the dedent
check. This prevents `:` from triggering auto indent when it's unwanted.