suggest/manager: treat project config as the search boundary
Previously, when traversing a project directory to look for a potential
project file, we just ignored the global config files. This was done
because we can't deduce the true project file from these configuration.

However the fact that that the project file cannot be outside of this
directory was overlooked, and cause issues with unconventional project
layout as with #29.