suggest/use: truncate the reference list on first response
Originally I thought that `def` will always come first, however after
some testing I found out that `use` responses might appear before
`def`.

This commit make it so that the reference list won't be cleared when a
`def` was first received, but instead cleared when the first response is
received.