indent/nim.vim: prevents infinite loop during search
The search for a closing parenthesis in the pragma indenter makes use of
the 'c' flag (accept match at cursor) to resolve the following:

  proc foo()
           ^ search for ')' starts here

Without the 'c' flag, the targeted parenthesis will be skipped, which is
unwanted.

However, this behavior is only correct iff only one search is necessary.

Consider the following:

  proc foo() {.importcpp: "foo(@)".}
                                ^ first search result

This ')' is inside a string, which means it is syntactically
insiginificant, so we retry for an another match.

However, since the 'c' flag is there, the next search yields at the same
location, causing an infinite loop.

This commit removes the 'c' flag from the list of flags after the first
match.

Of course, it is worth noting that the current deduced position is
completely wrong, since foo() already has a pragma in this case.
A fix for this will come later.

Fixes #41.